requirements.txt
pydantic==2.10.4
httpx==0.28.1
pillow==11.0.0
numpy==2.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
